Move code around pt 2
