+++ src/internal/m365/service/sharepoint/enabled_test.go
@@ -1,0 +1,102 @@
+package sharepoint
+
+import (
+	"context"
+	"testing"
+
+	"github.com/alcionai/clues"
+	"github.com/microsoftgraph/msgraph-sdk-go/models"
+	"github.com/microsoftgraph/msgraph-sdk-go/models/odataerrors"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/suite"
+
+	"github.com/alcionai/corso/src/internal/m365/graph"
+	"github.com/alcionai/corso/src/internal/tester"
+)
+
+type EnabledUnitSuite struct {
+	tester.Suite
+}
+
+func TestEnabledUnitSuite(t *testing.T) {
+	suite.Run(t, &EnabledUnitSuite{Suite: tester.NewUnitSuite(t)})
+}
+
+var _ getSiteRooter = mockGSR{}
+
+type mockGSR struct {
+	response models.Siteable
+	err      error
+}
+
+func (m mockGSR) GetRoot(context.Context) (models.Siteable, error) {
+	return m.response, m.err
+}
+
+func odErrMsg(code, message string) *odataerrors.ODataError {
+	odErr := odataerrors.NewODataError()
+	merr := odataerrors.NewMainError()
+	merr.SetCode(&code)
+	merr.SetMessage(&message)
+	odErr.SetErrorEscaped(merr)
+
+	return odErr
+}
+
+func (suite *EnabledUnitSuite) TestIsSharePointServiceEnabled() {
+	table := []struct {
+		name      string
+		mock      func(context.Context) getSiteRooter
+		expect    assert.BoolAssertionFunc
+		expectErr func(*testing.T, error)
+	}{
+		{
+			name: "ok",
+			mock: func(ctx context.Context) getSiteRooter {
+				return mockGSR{models.NewSite(), nil}
+			},
+			expect: assert.True,
+			expectErr: func(t *testing.T, err error) {
+				assert.NoError(t, err, clues.ToCore(err))
+			},
+		},
+		{
+			name: "no sharepoint license",
+			mock: func(ctx context.Context) getSiteRooter {
+				odErr := odErrMsg("code", string(graph.NoSPLicense))
+
+				return mockGSR{nil, graph.Stack(ctx, odErr)}
+			},
+			expect: assert.False,
+			expectErr: func(t *testing.T, err error) {
+				assert.NoError(t, err, clues.ToCore(err))
+			},
+		},
+		{
+			name: "arbitrary error",
+			mock: func(ctx context.Context) getSiteRooter {
+				odErr := odErrMsg("code", "message")
+
+				return mockGSR{nil, graph.Stack(ctx, odErr)}
+			},
+			expect: assert.False,
+			expectErr: func(t *testing.T, err error) {
+				assert.Error(t, err, clues.ToCore(err))
+			},
+		},
+	}
+	for _, test := range table {
+		suite.Run(test.name, func() {
+			t := suite.T()
+
+			ctx, flush := tester.NewContext(t)
+			defer flush()
+
+			gsr := test.mock(ctx)
+
+			ok, err := IsSharePointServiceEnabled(ctx, gsr, "resource_id")
+			test.expect(t, ok, "has sites flag")
+			test.expectErr(t, err)
+		})
+	}
+}
